[FIX]ConnexionError and remove canPop in ConfirmAppointment

See merge request misaina1/med_scheduler_front!6

+++ lib/Repository/MedecinRepository.dart
@@ -3,19 +3,15 @@ import 'package:http/http.dart' as http;
 import 'dart:async';
 import 'dart:convert';
 import 'package:med_scheduler_front/Utilisateur.dart';
-import 'package:med_scheduler_front/Specialite.dart';
 import 'package:med_scheduler_front/Medecin.dart';
 import 'package:med_scheduler_front/AuthProvider.dart';
 import 'package:med_scheduler_front/UrlBase.dart';
 import 'package:provider/provider.dart';
 import 'package:med_scheduler_front/main.dart';
-import 'package:med_scheduler_front/Centre.dart';
 import 'package:med_scheduler_front/CustomAppointment.dart';
 import 'package:med_scheduler_front/Utilitie/Utilities.dart';
-import 'package:med_scheduler_front/Espace_Admin/view/IndexAccueilAdmin.dart';
-import 'package:med_scheduler_front/Espace_Client/view/IndexAccueil.dart';
+import 'package:med_scheduler_front/ConnectionError.dart';
 import 'package:med_scheduler_front/Espace_Medecin/view/IndexAcceuilMedecin.dart';
-import 'package:jwt_decode/jwt_decode.dart';
 import 'package:intl/intl.dart';
 
 class MedecinRepository {
@@ -34,21 +30,21 @@ class MedecinRepository {
   Future<List<CustomAppointment>> getAllUnavalaibleAppointmentByDayAndDoctor(
       DateTime dtClicked, Medecin medecin,Medecin widgetMedecin) async {
 
+    if(await utilities.isConnectionAvailable()){
 
-    authProvider = Provider.of<AuthProvider>(context, listen: false);
-    token = authProvider.token;
-
-
-    String formated = DateFormat('yyyy-MM-dd').format(dtClicked);
+      authProvider = Provider.of<AuthProvider>(context, listen: false);
+      token = authProvider.token;
 
 
-    print('MED ID: ${widgetMedecin.id}');
-    final url = Uri.parse(
-        "${baseUrl}api/unavailable_appointments?page=1&startAt[before]=$formated&startAt[after]=$formated");
+      String formated = DateFormat('yyyy-MM-dd').format(dtClicked);
 
-    final headers = {'Authorization': 'Bearer $token'};
 
-    try {
+      print('MED ID: ${widgetMedecin.id}');
+      final url = Uri.parse(
+          "${baseUrl}api/unavailable_appointments?page=1&startAt[before]=$formated&startAt[after]=$formated");
+
+      final headers = {'Authorization': 'Bearer $token'};
+
       final response = await http.get(url, headers: headers);
 
       print('STATUS CODE APPOINTS AGENDA:  ${response.statusCode} \n');
@@ -78,32 +74,31 @@ class MedecinRepository {
         throw Exception(
             '-- Failed to load data. HTTP Status Code: ${response.statusCode}');
       }
-    } catch (e, stackTrace) {
-      if (e is http.ClientException) {
-        utilities.ErrorConnexion();
-      } else {
-        // Gérer d'autres exceptions
-        print('Une erreur inattendue s\'est produite: $e');
-      }
 
+    }else{
+
+      utilities.handleConnectionError(ConnectionError("Une erreur de connexion s'est produite!"));
       // Retourner une valeur par défaut en cas d'erreur
       return <CustomAppointment>[];
     }
+
+
   }
 
   Future<List<CustomAppointment>> getAllAppointmentMedecin(Medecin medecin) async {
 
-
-    authProvider = Provider.of<AuthProvider>(context, listen: false);
-    token = authProvider.token;
+    if(await utilities.isConnectionAvailable()){
 
 
-    final url = Uri.parse(
-        "${baseUrl}api/doctors/appointments/${utilities.extractLastNumber(medecin.id)}");
+      authProvider = Provider.of<AuthProvider>(context, listen: false);
+      token = authProvider.token;
 
-    final headers = {'Authorization': 'Bearer $token'};
 
-    try {
+      final url = Uri.parse(
+          "${baseUrl}api/doctors/appointments/${utilities.extractLastNumber(medecin.id)}");
+
+      final headers = {'Authorization': 'Bearer $token'};
+
       final response = await http.get(url, headers: headers);
 
       print('STATUS CODE APPOINTS AGNENDAAA: ${response.statusCode} \n');
@@ -135,32 +130,29 @@ class MedecinRepository {
         // Gestion des erreurs HTTP
         throw Exception('-- TOKEN EXPIRED.');
       }
-    } catch (e) {
-      if (e is http.ClientException) {
-        utilities.ErrorConnexion();
-      } else {
-        // Gérer d'autres exceptions
-        print('Une erreur inattendue s\'est produite: $e');
-      }
+    }else{
 
+      utilities.handleConnectionError(ConnectionError("Une erreur de connexion s'est produite!"));
       // Retourner une valeur par défaut en cas d'erreur
       return <CustomAppointment>[];
-
     }
+
   }
 
 
 
   Future<List<CustomAppointment>> getAllAppointment(Medecin medecin) async {
-    authProvider = Provider.of<AuthProvider>(context, listen: false);
-    token = authProvider.token;
 
-    final url = Uri.parse(
-        "${baseUrl}api/doctors/appointments/${utilities.extractLastNumber(medecin.id)}");
+    if(await utilities.isConnectionAvailable()){
 
-    final headers = {'Authorization': 'Bearer $token'};
+      authProvider = Provider.of<AuthProvider>(context, listen: false);
+      token = authProvider.token;
 
-    try {
+      final url = Uri.parse(
+          "${baseUrl}api/doctors/appointments/${utilities.extractLastNumber(medecin.id)}");
+
+      final headers = {'Authorization': 'Bearer $token'};
+
       final response = await http.get(url, headers: headers);
 
       if (response.statusCode == 200) {
@@ -180,17 +172,15 @@ class MedecinRepository {
         throw Exception(
             '-- Erreur d\'obtention des données\n vérifier votre connexion internet.');
       }
-    } catch (e, stackTrace) {
-      if (e is http.ClientException) {
-        utilities.ErrorConnexion();
-      } else {
-        // Gérer d'autres exceptions
-        print('Une erreur inattendue s\'est produite: $e');
-      }
 
+    }else{
+
+      utilities.handleConnectionError(ConnectionError("Une erreur de connexion s'est produite!"));
       // Retourner une valeur par défaut en cas d'erreur
       return <CustomAppointment>[];
     }
+
+
   }
 
 
@@ -200,10 +190,12 @@ class MedecinRepository {
       CustomAppointment appointment) async {
 
 
-    authProvider = Provider.of<AuthProvider>(context, listen: false);
-    token = authProvider.token;
+    if(await utilities.isConnectionAvailable()){
 
-    final url = Uri.parse("${baseUrl}api/appointments/${utilities!.extractLastNumber(appointment.id)}");
+      authProvider = Provider.of<AuthProvider>(context, listen: false);
+      token = authProvider.token;
+
+      final url = Uri.parse("${baseUrl}api/appointments/${utilities!.extractLastNumber(appointment.id)}");
     //final headers = {'Content-Type': 'application/json'};
 
     final headers = {
@@ -211,11 +203,9 @@ class MedecinRepository {
     'Authorization': 'Bearer $token'
     };
 
-    try {
-    print('PATCH ISDELETED: ${appointment.isDeleted}');
+
     String jsonUser = jsonEncode(appointment.toJsonUnav());
 
-    print('JSOON USER: ${jsonUser}');
 
     final response = await http.patch(url, headers: headers, body: jsonUser);
     print('${response.statusCode} \n BODY PATCH: ${response.body} ');
@@ -242,16 +232,10 @@ class MedecinRepository {
     'Il y a une erreur APPOINTMENT. HTTP Status Code: ${response.statusCode}');
     }
     }
-    } catch (e, stackTrace) {
-    if (e is http.ClientException) {
-    utilities.ErrorConnexion();
-    } else {
-    // Gérer d'autres exceptions
-    print('Une erreur inattendue s\'est produite: $e');
+    }else{
+    utilities.handleConnectionError(ConnectionError("Une erreur de connexion s'est produite!"));
     }
 
-    throw e;
-    }
   }
 
 
@@ -259,19 +243,20 @@ class MedecinRepository {
   Future<void> createUnavalaibleAppointment(
       CustomAppointment appointment) async {
 
+    if(await utilities.isConnectionAvailable()){
 
-    authProvider = Provider.of<AuthProvider>(context, listen: false);
-    token = authProvider.token;
+      authProvider = Provider.of<AuthProvider>(context, listen: false);
+      token = authProvider.token;
 
-    final url = Uri.parse("${baseUrl}api/unavailable_appointments");
-    //final headers = {'Content-Type': 'application/json'};
+      final url = Uri.parse("${baseUrl}api/unavailable_appointments");
+      //final headers = {'Content-Type': 'application/json'};
 
-    final headers = {
-      'Content-Type': 'application/ld+json',
-      'Authorization': 'Bearer $token'
-    };
+      final headers = {
+        'Content-Type': 'application/ld+json',
+        'Authorization': 'Bearer $token'
+      };
 
-    try {
+
       String jsonUser = jsonEncode(appointment.toJsonUnav());
 
       final response = await http.post(url, headers: headers, body: jsonUser);
@@ -299,16 +284,9 @@ class MedecinRepository {
               'Il y a une erreur APPOINTMENT. HTTP Status Code: ${response.statusCode}');
         }
       }
-    } catch (e, stackTrace) {
-      if (e is http.ClientException) {
-        utilities.ErrorConnexion();
-      } else {
-        // Gérer d'autres exceptions
-        print('Une erreur inattendue s\'est produite: $e');
-      }
 
-
-      throw e;
+    }else{
+      utilities.handleConnectionError(ConnectionError("Une erreur de connexion s'est produite!"));
     }
   }
 
@@ -316,68 +294,65 @@ class MedecinRepository {
       CustomAppointment appointment) async {
 
 
-    authProvider = Provider.of<AuthProvider>(context, listen: false);
-    token = authProvider.token;
+    if(await utilities.isConnectionAvailable()){
 
-    final url = Uri.parse(
-        "${baseUrl}api/unavailable_appointments/${utilities.extractLastNumber(appointment.id)}");
-    //final headers = {'Content-Type': 'application/json'};
+      authProvider = Provider.of<AuthProvider>(context, listen: false);
+      token = authProvider.token;
 
-    final headers = {'Authorization': 'Bearer $token'};
+      final url = Uri.parse(
+          "${baseUrl}api/unavailable_appointments/${utilities.extractLastNumber(appointment.id)}");
+      //final headers = {'Content-Type': 'application/json'};
 
-    try {
-    //print('Request Body: $jsonUser');
-    final response = await http.delete(url, headers: headers);
-    print(response.statusCode);
+      final headers = {'Authorization': 'Bearer $token'};
 
-    if (response.statusCode == 200) {
-    final Map<String, dynamic> jsonResponse = json.decode(response.body);
 
-    if (jsonResponse.containsKey('error')) {
-    utilities.error('Rendez-vous déja existant');
-    } else {}
-    } else {
+      final response = await http.delete(url, headers: headers);
+      print(response.statusCode);
 
-    if (response.statusCode == 401) {
-    authProvider.logout();
-    Navigator.pushReplacement(
-    context, MaterialPageRoute(builder: (context) => const MyApp()));
-    }
-    if (response.statusCode == 204) {
-    } else {
-    // Gestion des erreurs HTTP
+      if (response.statusCode == 200) {
+        final Map<String, dynamic> jsonResponse = json.decode(response.body);
 
-    utilities.error(
-    'Il y a une erreur APPOINTMENT. HTTP Status Code: ${response.statusCode}');
-    }
-    }
-    } catch (e, stackTrace) {
-      if (e is http.ClientException) {
-        utilities.ErrorConnexion();
+        if (jsonResponse.containsKey('error')) {
+          utilities.error('Rendez-vous déja existant');
+        } else {}
       } else {
-        // Gérer d'autres exceptions
-        print('Une erreur inattendue s\'est produite: $e');
-      }
 
-    throw e;
+        if (response.statusCode == 401) {
+          authProvider.logout();
+          Navigator.pushReplacement(
+              context, MaterialPageRoute(builder: (context) => const MyApp()));
+        }
+        if (response.statusCode == 204) {
+        } else {
+          // Gestion des erreurs HTTP
+
+          utilities.error(
+              'Il y a une erreur APPOINTMENT. HTTP Status Code: ${response.statusCode}');
+        }
+      }
+    }else{
+      utilities.handleConnectionError(ConnectionError("Une erreur de connexion s'est produite!"));
     }
+
   }
 
 
 
   Future<List<CustomAppointment>> getAllAppointmentByMedecin(
       Utilisateur user) async {
-    authProvider = Provider.of<AuthProvider>(context, listen: false);
-    token = authProvider.token;
 
-    print('ID MED: ${user.id}');
+    if(await utilities.isConnectionAvailable()){
 
-    final url = Uri.parse(
-        "${baseUrl}api/doctors/appointments/${utilities.extractLastNumber(user.id)}");
+      authProvider = Provider.of<AuthProvider>(context, listen: false);
+      token = authProvider.token;
 
-    final headers = {'Authorization': 'Bearer $token'};
+      print('ID MED: ${user.id}');
 
-    try {
+      final url = Uri.parse(
+          "${baseUrl}api/doctors/appointments/${utilities.extractLastNumber(user.id)}");
+
+      final headers = {'Authorization': 'Bearer $token'};
+
       final response = await http.get(url, headers: headers);
 
       print('STATUS CODE APPOINTS: ${response.statusCode} \n');
@@ -400,32 +375,29 @@ class MedecinRepository {
         throw Exception(
             '-- Erreur d\'obtention des données\n vérifier votre connexion internet.');
       }
-    } catch (e) {
-      if (e is http.ClientException) {
-        utilities.ErrorConnexion();
-      } else {
-        // Gérer d'autres exceptions
-        print('Une erreur inattendue s\'est produite: $e');
-      }
+    }else{
 
+      utilities.handleConnectionError(ConnectionError("Une erreur de connexion s'est produite!"));
       // Retourner une valeur par défaut en cas d'erreur
       return <CustomAppointment>[];
     }
+
   }
 
 
 
   Future<void> getUserToChangePassword(int id,String newPassword) async {
-    final url = Uri.parse("${baseUrl}api/change-password/$id");
 
-    print('URL USER: $url');
+    if(await utilities.isConnectionAvailable()){
 
-    //final headers = {'Authorization': 'Bearer $token'};
+      final url = Uri.parse("${baseUrl}api/change-password/$id");
 
-    final body = {"password": "$newPassword"};
+      print('URL USER: $url');
 
+      //final headers = {'Authorization': 'Bearer $token'};
 
-    try {
+      final body = {"password": "$newPassword"};
+
       final response = await http.patch(url, body: jsonEncode(body));
       print(' --- ST CODE: ${response.statusCode}');
 
@@ -445,15 +417,12 @@ class MedecinRepository {
         }
         throw Exception('ANOTHER ERROR');
       }
-    } catch (e, stackTrace) {
-      if (e is http.ClientException) {
-        utilities.ErrorConnexion();
-      } else {
-        // Gérer d'autres exceptions
-        print('Une erreur inattendue s\'est produite: $e');
-      }
-      throw Exception('-- Failed to load data. Error: $e');
+
+    }else{
+
+      utilities.handleConnectionError(ConnectionError("Une erreur de connexion s'est produite!"));
     }
+
   }
